pyproject.toml
[tool.poetry]
name = "feed-service"
version = "0.1.0"
description = ""
authors = ["JungJinHwan <jinhjung@cuneo.co.kr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
djangorestframework = "^3.14.0"
ipython = "^8.18.0"


[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
pytest = "^7.4.3"
pytest-django = "^4.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.unittest"
python_files = ["test_*.py", "*_test.py"]
